requirements.txt
asgiref==3.9.1
Django==5.2.6
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
setuptools==78.1.1
sqlparse==0.5.3
wheel==0.45.1
whitenoise==6.9.0
python-dateutil==2.9.0
Pillow>=10.0.0
django-crontab==0.7.1
django-cors-headers==4.6.0
python-dotenv==1.0.1
